update nav panels
